Refactored Dispatcher lib, it now takes a Router_Url instance
Dispatcher lib no longer takes data directly from the main router, instead
Dispatcher::dispatch() takes a Router_Url instance which it takes the data from.

It also no longer loads a 'NPC', this is instead done in the Bootstrap which
falls back to the 'session' module if the use does not have permission.

Dropped support for full page errors in the Dispatcher, it should not be its job
to do this.

Some things are broke in this commit, e.g. no controller title, no page links

- Removed Dispatcher::isStandalone() Dispatcher::atFrontpage()
  Dispatcher::error()
- Dispatcher::displayErrors() renamed to Dispatcher::setDisplayErrors()
- Dispatcher::getReqCntrl() renamed to Dispatcher::getReqCntrlr()
- Added Dispatcher::setStatusHeader() to set HTTP status code
- Dispatcher::dispatch() now has 2 args, $request (Router_Url) and array $config
- Tweaked 404 and 403 view files, more human friendly
- Hook bootstrap_loaded now has 2 additional args, Dispatcher status code and
  dispatch data