Implemented 'users_meta' table (feature #263)
The 'users' table has been split into two, with the second being a new
'users_meta' table. This will hold all optional/additional details of a user,
such as Birthday, Sex, Location - what ever a module wants.

When adding/editing a user, anything that is not a 'standard user key' will be
placed within the 'users_meta' table. To remove a meta detail, simply edit the
user and provide NULL as the value for the key.

- Ugmanager::getUser() added 3rd argument to get data from 'users_meta' table,
  bool false by default
- Added Ugmanager::getUserMetaData()
- Removed Ugmanager::awaitingValidation(), Ugmanager::activateUser() and
  Ugmanager::resetPassword() since this should not be a feature of Ugmanager
- Removed Ugmanager_InvalidResetCode and Ugmanager_InvalidActiviationCode